Move more functionality out of `da.core`
- move tile, repeat, fromfunction to da.creation
- delete duplicate `same_keys` definitions.
