+++ dask/array/core.py
@@ -1795,6 +1795,7 @@ class Array(Base):
 
     @wraps(np.repeat)
     def repeat(self, repeats, axis=None):
+        from .creation import repeat
         return repeat(self, repeats, axis=axis)
 
     @wraps(np.nonzero)
@@ -3127,24 +3128,6 @@ def offset_func(func, offset, *args):
     return _offset
 
 
-@wraps(np.fromfunction)
-def fromfunction(func, chunks=None, shape=None, dtype=None):
-    if chunks:
-        chunks = normalize_chunks(chunks, shape)
-    name = 'fromfunction-' + tokenize(func, chunks, shape, dtype)
-    keys = list(product([name], *[range(len(bd)) for bd in chunks]))
-    aggdims = [list(accumulate(add, (0,) + bd[:-1])) for bd in chunks]
-    offsets = list(product(*aggdims))
-    shapes = list(product(*chunks))
-
-    values = [(np.fromfunction, offset_func(func, offset), shp)
-              for offset, shp in zip(offsets, shapes)]
-
-    dsk = dict(zip(keys, values))
-
-    return Array(dsk, name, chunks, dtype=dtype)
-
-
 @wraps(np.unique)
 def unique(x):
     name = 'unique-' + x.name
@@ -3873,67 +3856,6 @@ def swapaxes(a, axis1, axis2):
 
     return atop(np.swapaxes, out, a, ind, axis1=axis1, axis2=axis2,
                 dtype=a.dtype)
-
-
-@wraps(np.repeat)
-def repeat(a, repeats, axis=None):
-    if axis is None:
-        if a.ndim == 1:
-            axis = 0
-        else:
-            raise NotImplementedError("Must supply an integer axis value")
-
-    if not isinstance(repeats, Integral):
-        raise NotImplementedError("Only integer valued repeats supported")
-
-    if -a.ndim <= axis < 0:
-        axis += a.ndim
-    elif not 0 <= axis <= a.ndim - 1:
-        raise ValueError("axis(=%d) out of bounds" % axis)
-
-    if repeats == 1:
-        return a
-
-    cchunks = np.cumsum((0,) + a.chunks[axis])
-    slices = []
-    for c_start, c_stop in sliding_window(2, cchunks):
-        ls = np.linspace(c_start, c_stop, repeats).round(0)
-        for ls_start, ls_stop in sliding_window(2, ls):
-            if ls_start != ls_stop:
-                slices.append(slice(ls_start, ls_stop))
-
-    all_slice = slice(None, None, None)
-    slices = [(all_slice,) * axis + (s,) + (all_slice,) * (a.ndim - axis - 1)
-              for s in slices]
-
-    slabs = [a[slc] for slc in slices]
-
-    out = []
-    for slab in slabs:
-        chunks = list(slab.chunks)
-        assert len(chunks[axis]) == 1
-        chunks[axis] = (chunks[axis][0] * repeats,)
-        chunks = tuple(chunks)
-        result = slab.map_blocks(np.repeat, repeats, axis=axis, chunks=chunks,
-                                 dtype=slab.dtype)
-        out.append(result)
-
-    return concatenate(out, axis=axis)
-
-
-@wraps(np.tile)
-def tile(A, reps):
-    if not isinstance(reps, Integral):
-        raise NotImplementedError("Only integer valued `reps` supported.")
-
-    if reps < 0:
-        raise ValueError("Negative `reps` are not allowed.")
-    elif reps == 0:
-        return A[..., :0]
-    elif reps == 1:
-        return A
-
-    return concatenate(reps * [A], axis=-1)
 
 
 def slice_with_dask_array(x, index):
